requirements.txt
pymongo
bcrypt
flask
pyjwt
python-dotenv